move out requantification_handler.pyi
